Move `PlayfairEncoder` to separate module
